feat: 콘솔, 웹 중복코드 제거
