+++ README.md
@@ -32,7 +32,7 @@
         - name
         - position
 - [x] 기존 기능 수정 - 출력 방식 수정
-- [ ] 리팩터링 - 중복 코드 제거
+- [x] 리팩터링 - 중복 코드 제거
 
 ## 추가 기능
